backend-fastapi/requirements.txt
fastapi
uvicorn
firebase-admin
google-cloud-vision
requests
python-multipart
